refactor: 법안 카테고리와 정당 헬퍼 모듈 통합
